requirements.txt
mcp>=1.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
python-pptx>=1.0.0
